+++ core/src/test/kotlin/net/corda/core/flows/TestDataVendingFlow.kt
@@ -1,13 +1,12 @@
-package net.corda.testing
+package net.corda.core.flows
 
 import co.paralleluniverse.fibers.Suspendable
-import net.corda.core.flows.SendStateAndRefFlow
 import net.corda.core.identity.Party
 import net.corda.core.internal.FetchDataFlow
 import net.corda.core.utilities.UntrustworthyData
 
 // Flow to start data vending without sending transaction. For testing only.
-class DataVendingFlow(otherSide: Party) : SendStateAndRefFlow(otherSide, emptyList()) {
+class TestDataVendingFlow(otherSide: Party) : SendStateAndRefFlow(otherSide, emptyList()) {
     @Suspendable
     override fun sendPayloadAndReceiveDataRequest(otherSide: Party, payload: Any): UntrustworthyData<FetchDataFlow.Request> {
         return if (payload is List<*> && payload.isEmpty()) {
